Started the post-processing-consumer service.
The consumer is configured to work with buses, but right now we do not have buses so it's commented out.
It can either handle all jobs or just specific one job.
Also, started implementing the MetadaraParserService.
Right now it knows to parse a .data file of TekenHozi 1.0 into objects, and the next task is to index them in ElasticSearch.